core: Move where_clause.rs to translate/
